profiles/desktop: set the USE_EXPAND variables in the profile defaults
It seems that we need to specify which USE_EXPAND variables are set in
the profile's make.defaults file, otherwise portage seems to ignore
the variable. These have been update to match the current as-built
configuration.

We can now delete the configuration from make.conf files on all systems
using the desktop profile.