profiles: Fixes profiles for changes to gentoo's licence policy
I would prefer if I could define licenses for specific packages for the
profiles in package.license, but that is not supported by portage (per man
portage), so we are forced to enable them globally across all packages here.

License groups for profiles in this repo are defined in
profiles/license_groups and pulled in by settings in the profiles'
make.default files.