profiles/common: Fixes the SYMLINK_LIB mangling in profile.bashrc
The fall-through case was to assume that the profile has been migrated
and set the SYMLINK_LIB and LIBDIR_x86 variables according to the 17.1
profile. That fails when installing into a empty rootfs and the
profile buildroot has not been migrated to a 17.1 profile.

We now check for the existance of a non-symlinked /lib directory
before assuming a migrated profile, so the fall-through case is a
no-op and so it is dependent on the profiles parent configuration.