profiles/desktop: Fix gio-launch-desktop breakage
dev-libs/glib-2.64.2 removed gio-launch-desktop. It is still needed by
mate-panel for loading .desktop files from the menu and panel.

This change creates a bashrc postinst action to create a simple script
to replace the deleted application helper when dev-lib/glib is next
installed. This is a pretty ugly was to fix the issue, to be honest,
but I hope that mate-panel will clean up this mess soon.